Move ReleaseCard to a component
